ENH: OBJWriter-Improve precision of saved points, normals and tcoords
OBJWriter saves points, normals and vectors in ASCII. This can limit
the precision of the written coordinates. When they are read using
vtkOBJWriter, the coordinates may not match the original values.

This MR uses vtkNumberToString to convert the floating point values
to high precision. When they are read back, there is 0 difference in
the
original values.

TestOBJPolyDataWriter was changed to set the tolerance for error to 0.