Use actual ValueTypes when computing value ranges.
Fixes #17664.

We currently truncate to double precision, which gives incorrect results
on 64-bit integers.

We need to enable `/bigobj` on MSVC to support the larger symbol tables
needed to handle the template codegen.

This does not cache the ranges like GetRange does, since we do not have
the infrastructure to do so (we don't have information keys for large
integers). Issue #17666 reflects this.

Since the GetRange implementations in vtkDataArrayPrivate.txx
depend on vtkSMPTools, which pulls in tbb, which pulls in
windows.h under MSVC, we cannot directly include this code
in the GenericDataArray headers. Instead, 64-bit integer
implementations are compiled into a new vtkGenericDataArray.cxx
translation unit for AOS, SOA, and ScaledSOA data arrays. All
other array/value types will go through the double interface
of vtkDataArray::ComputeRange.

A test has been added to ensure that the common array types
are able to return their full ValueType range from
GetValueRange to make sure we continue to support the full
ranges going forward.