Fix crash on axis cut filter for hyper tree grid
In the case of slicing on cell boundaries, the filter added children
from 2 different cells when it should only do it for one.
Changing the test condition for going through the children fixes the
problem.

This commit also added a call to Initialize from vtkHyperTreeGrid
at the beginning of the filter to make sure we are with clean data