Fix vtkThinPlateSplineTransform forward transform for coplanar points
vtkThinPlateSplineTransform provided incorrect output when all the source
points were coplanar and/or target points were coplanar and one of the planes
were exactly aligned with coordinate system axes. Root cause was that
in this case the A linear transformation matrix in
vtkThinPlateSplineTransform::InternalUpdate was singular and it was
"regularized" by setting it to identity.

Added option to enable/disable bulk transform regularization, thereby
allowing forward transform to be computed correctly in all configurations.
Bulk transform regularization is still useful when all source and target
points are in XY plane because it allows computation of the inverse
transform.

Added "TestThinPlateSplineTransform" to test the transform in various
conditions (dimensionality, noise, rotation, translation, etc).