Fixing Bugs In CIEDE2000 ColorInterpolation Method
This fixes the steep color gradients appearing when using the Lab2000
color interpolation method. Also an option was added to force the support
colors to be exact as opposed to being snapped to grid. Since the visual
results are better if thay snap to grid, the option is set to false by
default.

Adding Reference Of Algorithm Publication