Adding an AddPointItem to vtkControlPointsItem
Previous implementation of vtkControlPointsItem was
problematic as it "Hit" everywhere in order to be able to
add points. This adds a new AddPointItem that can be recovered
from any vtkControlPointsItem
When enabled, the vtkControlPointsItem hit only on the actual points
while the AddPointItem hit everywhere.
By placing the AddPointItem below all other item and the vtkControlPointItem above all other items,
other items can be interacted with while keeping interaction with the vtkControlPointItem.