let pathtracers use either backplate or environmental light or both
The backplate is like standard rendering's background. It does not
move with the camera and it does not affect the appearance of objects
in the scene.

The environmental map moves with the camera and affects the appearance
of objects in the scene. It is like VTK's newer SkyBox, with material
effects turned on.

SetBackgroundMode let's you choose which behaviour, or both. The case
of using neither (to get a translucent background) is setable but
has not been proven out in practice yet.

The environment mode controls (color, gradient, texture) have new
methods on vtkRenderer but these have no effect on standard rendering.
Likewise SkyBox mode is not respected by the pathtracer. Unifying both
of these has yet to be done.

The default for the pathtracer was changed to use backplate and disable
environment lighting. This makes default object brightness similar
to standard VTK.