Update of htg clipping tests + threshold filter
vtkThreshold thresholding methods were inlined. Since
the method used (Lower, Upper, or Between) is stored
as a function pointer in ThresholdFunction, the address
of the method was different in children. Removing the "inlinement"
removes this flaw. The performance is unaletered because the
threshold estimation was already performed through a function
pointer that was not inlined.