Update vtkMarchingSquares to use dispatch/range.
The diff is noisy, but the change to the implemenation was trivial --
just replaced `T *scalars` with a `DataArrayValueRange` and the rest
of the code works as before.

I removed the special case for multiple components, since it just fed
the multi-component array into the single-component implementation. Not
sure what it was supposed to do.