Reader for temporal CSV.
Read a CSV file and interpret a numerical column as time indicator.

An internal reader is used to read the csv file. This reader behave like
the vtkDelimitedTextReader, but have the SetTimeColumn Name/Id in
addition to set the time indicator. It expose the FileName,
FieldDelimiterCharacters, HaveHeaders, MergeConsecutiveDelimiters and
AddTabFieldDelimiter parameters for convenience (also exposed in
paraview). If more parameters are required for the CSV Reader, it is
possible the get the internal reader.