Added texture writing to vtkJSONSceneExporter
This includes a simple JPEG texture writer that saves the texture
inside the .vtkjs file, and a texture LOD series writer that saves
the textures as a series of decreasing resolution jpeg files intended
to be uploaded to the web. The texture LOD files are saved outside of
the .vtkjs file.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>