Add poly LOD writing to vtkJSONSceneExporter
This adds options to write polyLODs in a series of decreasing
resolution directories. The directories do not get zipped here, but
are intended to be zipped in subclasses, including the vtkPVWebExporter
in ParaView. The zipped files are intended to be uploaded to the web,
and with some new additions to VTK.js, they will be downloaded one at
a time, which has the effect of increasing the resolution of the polydata
in VTK.js over time.

For the quadric clustering, the initial guess for the number of
divisions may be improved upon, as well as some of the math involved.
The goal is for each directory to be roughly 1/4 the size of the
previous one, but it is not always straightforward to do when using the
vtkQuadricClustering filter.

Signed-off-by: Patrick Avery <patrick.avery@kitware.com>