Add scene graph-enabled exporter for vtk-js
This scene exporter differs from vtkJSONSceneExporter by exposing more
class-to-class parity between VTK and vtk-js. It also utilizes the
rendering module's scene graph traversal routines, which allow for the
export of more than one renderer associated with a render
window. Finally, it provides a layer of abstraction between the output
data's format and its transcription, facilitating the use of
compression, archival and in-memory serialization.