FindHDF5 Bugfix For Linking Against Correct External HDF5
Setting up VTK to use an external HDF5 interactively, instead of passing
all necessary variables at once, causes the VTK libraries to be linked
against an external HDF5 in a standard location.
This happens despite FindHDF5 successfully finding the desired external
HDF5 in the non-standard location.
Changing the precedence in the if statements handling the location
assignment for the import targets resovles this issue.