Change the default behavior back to what it was before.
This will keep the optix dashboard happy and is less of a
change to users. Note the distinction between BP and ENV
though, means that code changes are still necessary to
get the same pixels out, as in the test changes here, several
of which exercise ray cast and path trace.