vtkExtractCells fixes
vtkExtractCells changed the cell ids filled by the user in its
RequestData. This was incorrect. It had the potential of dropping cell
ids provided is the input dataset had fewer elements for an earlier
iteration.

Adding API to vtkExtractCells to skip sorting and uniquifying operations
if its already known that the ids are sorted and unique.