vtkXMLMultiBlockDataReader: fix when reading nested multipiece datasets
When a vtkMultiBlockDataSet contains a vtkMultiPieceDataSet,
vtkXMLPMultiBlockDataWriter will split each piece into another
vtkMultiPieceDataSet. As a result, the written XML data file will
contain nested "Piece" elements. Since vtkMultiPieceDataSets are not
hierarchical, this dataset cannot be recreated as described.

To fix, have the reader replace the top-level vtkMultiPieceDataSet
with a vtkMultiBlockDataSet instead when nested Piece elements are
found.

Modified test to exercise writing out a vtkMultiPieceDataSet block and
read it back in.