Fix anomalous spikes in contour bands output
The vtkBandedPolyDataContourFilter::ClipEdge method interpolates between
two edge points, creating extra points at the intermediate clip values,
if any. The method retrieves the clip values for the edge points. The
clip value for the highest scalar value may be somewhat larger than the
end point scalar value due to the application of the internal clip
tolerance. When the difference between the end point values is on the
order of this clip tolerance this may result in an interpolation factor
significantly larger than 1. The effect of this overshoot is that spikes
appear extending outside the original cell.

This issue is fixed by excluding the high end point clip value if the
resulting interpolation factor is significantly larger than 1.