Move interactors into new UI module
These interactors are based on their native
UI but do not have any requirement on OpenGL.
Move them into this new directory so that Vulkan
or other toolkits can make use of them.

While at it remove the use of Xt as it is no longer
really used and instead simplify the code and go directly
to Xlib.