Update vtkExtractVOI filter to respect vtkImage::DirectionMatrix
An oriented image needs to have its direction matrix passed to the
output, and the boundaries and origin need to respect the
direction matrix.

Change the filter's test to use and oriented input, and check for
the direction matrix in the output.

Fix compliation error for debug VS2019 build.
VC compiler doesn't recognize the 'not' keyword, only '!'.