Apply additional hierarchy files when wrapping
The module hierarchy file includes all the classes that are
implemented in the DEPENDS dependency chain.  The wrappers,
however, also require information about certain classes that
are implemented outside of this dependency chain.  Specifically,
if a class header contains a forward reference to another class,
and if that other class is used in the interface, then the
wrappers require a hierarchy file that provides information about
that class.

A solution to this problem is for the wrappers to use the
hierarchy files for all modules in PRIVATE_DEPENDS, in addition
to using the module's own hierarchy file.  This will bring in
more hierarchy files than is strictly necessary, of course,
but too many is better than too few.