Fixed incorrect GetScreenSize implementation in Cocoa, Win32, and OpenVR
Cocoa amd Win32 were both writing to the Size ivar, which is not supposed to be the size of the entire screen.

The Win32 implementation had a ScreenSize ivar but it was totally unused.  The Cocoa implementation didn't even have a ScreenSize ivar.  Put ScreenSize ivar up into vtkOpenGLRenderWindow, and removed it from subclasses that had such an ivar. The OpenVR implementation was missing (and provided to me by Ken).

Also, in Cocoa, instead of getting the NSScreen from the NSWindow, go through the NSView, which is the thing VTK is drawing into, and so slightly more correct.

Added TODOs for subclasses that don't have any code to determine screen sizes.