Fix infinite loop in vtkContourTriangulator
During vtkCCSCutHoleyPolys, the same polys could be specified as both the outer and inner contours.
This would cause an infinite loop as polys were repeatedly added and removed from the list, and as a result the loop would never be terminated.

Fixed by adding checks to prevent the same poly from being specified as both inner and outer contours.

Fixes: #17627